// packages/components/src/data-table/src/interface.ts
export type FilterOptionValue = string | number
export type ColumnKey = string | number
export type RowKey = string | number

export type SortOrderFlag = 1 | -1 | 0

export type RowData = Record<string, any>

export type InternalRowData = Record<string, unknown>
// ...
export type Sorter<T = InternalRowData> = CompareFn<T> | SorterMultiple<T>
export interface SorterMultiple<T = InternalRowData> {
  multiple: number
  compare?: CompareFn<T> | 'default'
}

export type Filter<T = InternalRowData> = (
  filterOptionValue: FilterOptionValue,
  row: T
) => boolean

export interface FilterOption {
  label: string
  value: FilterOptionValue
}

export type TmNode = TreeNode<InternalRowData>

// for compat may add null
export type SortOrder = 'ascend' | 'descend' | false

export type Ellipsis = boolean | (EllipsisProps & { style?: CSSProperties })

export interface CommonColumnInfo<T = InternalRowData> {
  fixed?: 'left' | 'right'
  width?: number | string
  minWidth?: number | string
  maxWidth?: number | string
  className?: string
  align?: 'left' | 'center' | 'right'
  titleAlign?: 'left' | 'center' | 'right'
  ellipsis?: Ellipsis
  cellProps?: (rowData: T, rowIndex: number) => HTMLAttributes
  description?: string | ((column: TableColumn<T>) => VNodeChild)
  icon?: () => VNodeChild
}

export type TableColumnTitle =
  | string
  | ((column: TableBaseColumn) => VNodeChild)

export type TableExpandColumnTitle =
  | string
  | ((column: TableExpandColumn) => VNodeChild)

export type TableColumnGroupTitle =
  | string
  | ((column: TableColumnGroup) => VNodeChild)

export type TableColumnGroup<T = InternalRowData> = {
  title?: TableColumnGroupTitle
  type?: never
  key: ColumnKey
  children: Array<TableBaseColumn<T>>

  // to suppress type error in table header
  visible?: boolean
  resizable?: boolean
  filterOptions?: never
  description?: string | ((column: TableBaseColumn) => VNodeChild)
  icon?: () => VNodeChild
} & CommonColumnInfo<T>

export type TableBaseColumn<T = InternalRowData> = {
  title?: TableColumnTitle
  titleColSpan?: number
  // for compat maybe default
  type?: never
  key: ColumnKey

  tree?: boolean

  visible?: boolean

  sorter?: boolean | Sorter<T> | 'default'
  defaultSortOrder?: SortOrder
  sortOrder?: SortOrder // controlled

  resizable?: boolean
  minWidth?: string | number
  maxWidth?: string | number

  filter?: 'default' | boolean | Filter<T>
  filterOptions?: FilterOption[]
  filterOptionValues?: FilterOptionValue[] | null // controlled
  filterOptionValue?: FilterOptionValue | null // controlled
  filterMode?: 'or' | 'and'

  defaultFilterOptionValues?: FilterOptionValue[] | null
  defaultFilterOptionValue?: FilterOptionValue | null
  filterMultiple?: boolean

  render?: (rowData: T, rowIndex: number) => VNodeChild
  renderFilter?: RenderFilter
  renderFilterIcon?: RenderFilterIcon
  renderSorter?: RenderSorter
  renderSorterIcon?: RenderSorterIcon
  renderFilterMenu?: RenderFilterMenu
  colSpan?: (rowData: T, rowIndex: number) => number
  rowSpan?: (rowData: T, rowIndex: number) => number
  description?: string | ((column: TableColumn<T>) => VNodeChild)
  icon?: () => VNodeChild
} & CommonColumnInfo<T>

export type TableSelectionColumn<T = InternalRowData> = {
  type: 'selection'
  multiple?: boolean
  disabled?: (row: T) => boolean
  options?: DataTableSelectionOptions<T>

  // to suppress type error in utils
  key?: ColumnKey
  title?: string | TableExpandColumnTitle
  visible?: boolean
  sorter?: never
  resizable?: boolean
  filter?: never
  filterOptions?: never
  filterOptionValues?: never
  filterOptionValue?: never
  colSpan?: never
  rowSpan?: never
  description?: string | ((column: TableColumn<T>) => VNodeChild)
  icon?: () => VNodeChild
} & CommonColumnInfo<T>

export type RenderExpand<T = InternalRowData> = (
  row: T,
  index: number
) => VNodeChild
export type RenderExpandIcon = ({
  expanded
}: {
  expanded: boolean
}) => VNodeChild

// TODO: we should deprecate `index` since it would change after row is expanded
export type Expandable<T = InternalRowData> = (row: T) => boolean
export interface TableExpandColumn<T = InternalRowData>
  extends Omit<TableSelectionColumn<T>, 'type'> {
  type: 'expand'
  title?: TableExpandColumnTitle
  renderExpand: RenderExpand<T>
  expandable?: Expandable<T>

  //
  key?: ColumnKey
}

export type TableColumn<T = InternalRowData> =
  | TableColumnGroup<T>
  | TableBaseColumn<T>
  | TableSelectionColumn<T>
  | TableExpandColumn<T>
export type TableColumns<T = InternalRowData> = Array<TableColumn<T>>

export type DataTableSelectionOptions<T = InternalRowData> = Array<
| DataTableSelectionOption
| { label: string, key: string | number, onSelect: (pageData: T[]) => void }
>
// ...
export type RenderFilter = (props: {
  active: boolean
  show: boolean
}) => VNodeChild

export type RenderFilterIcon = RenderFilter

export type RenderSorter = (props: { order: SortOrder }) => VNodeChild

export type RenderSorterIcon = RenderSorter

export type RenderFilterMenu = (actions: { hide: () => void }) => VNodeChild